commit so I can pull

+++ IonWeb/dispenser/views.py
@@ -10,8 +10,9 @@ from django.shortcuts import render
 import re
 
 def management(request):
+   params = {'dispensers': dispenser.objects.order_by('location'), 'compartments': compartment.objects}
+   
    if request.method == 'POST':
-
       if request.POST['requestType'] == 'removeCompartment':
          compID = request.POST['compID']
          dispID = request.POST['dispID']
@@ -27,16 +28,22 @@ def management(request):
       
       if request.POST['requestType'] == 'deleteDispenser':
          id = request.POST['id']
-         Dispenser = dispenser.objects(id=id)[0]
+         if dispenser.objects(id=id):
+            Dispenser = dispenser.objects(id=id)[0]
+            for Compartment in Dispenser.slots:
+               if Compartment:
+                  Compartment.loaded = False
+                  Compartment.save()
 
-         for Compartment in Dispenser.slots:
-            if Compartment:
-               Compartment.loaded = False
-               Compartment.save()
+               Dispenser.user.user.delete()
+               Dispenser.delete()            
+         else:
+            params['message'] = "Page refresh - not a valid dispenser to delete"
 
-         Dispenser.delete()
-
-   return render(request, 'dispenser.html', {'dispensers': dispenser.objects.order_by('location'), 'compartments': compartment.objects})
+      if request.POST['requestType'] == 'loadCompartment':
+         dispID = request.POST['dispID']
+         
+   return render(request, 'dispenser.html', params)
 
 def compartments(request):
    params = {}
@@ -180,7 +187,7 @@ def dispenser_view(request):
                               if int(rxuid) == Compartment.rxuid:
                                  validMedications[rxuid] = index, Patient, True
                   params['validMedications'] = validMedications
-                  
+                  params['caretaker'] = True
             elif ionuser.group == "patient":
                Patient = patient.objects(user=ionuser.id)[0]
       
@@ -250,3 +257,62 @@ def dispenser_admin(request):
 @is_in_group(DISPENSER_GROUP)   
 def dispense_medication(request):
   return render(request, 'dispense_medication.html')
+
+def load_compartment(request):
+   params = {}
+   # get dispenserID from user ID
+   if 'dispenserID' in request.GET:
+      dispID = request.GET['dispenserID']
+      if is_valid(dispID) and dispenser.objects(id=dispID):
+         Dispenser = dispenser.objects(id=dispID)[0]
+         params['dispenser'] = Dispenser
+      else:      
+         params['message'] = 'Invalid dispenser ID'
+         params['hide'] = True
+      
+      free = []
+      for index, slot in enumerate(Dispenser.slots):
+         if slot == None:
+            free.append(index)
+      params['free'] = free
+      
+      if len(free) == 0:
+         params['message'] = "No free slots in this dispenser."
+         params['hide'] = True
+         
+   if request.method == 'POST':
+      if request.POST['requestType'] == 'compID':
+         Dispenser = dispenser.objects(id=request.POST.get('dispID'))[0]
+         params['dispID'] = Dispenser.id
+         compID = request.POST.get('compID')
+         if is_valid(compID) and compartment.objects(id=compID):
+            Compartment = compartment.objects(id=compID)[0]
+            
+            if not Compartment.rxuid:
+               params['message'] = "Empty Compartment."
+            
+            elif Compartment.loaded == True:
+               location = helper.findCompartment(compID)
+               params['message'] = "Compartment already loaded in slot " + str((location[1] + 1)) + " in " + location[0].location + " dispenser."
+            else:
+               params['compID'] = compID
+         else:
+            params['message'] = "Not a valid compartment ID"
+      
+      # detect which slot?
+      if request.POST['requestType'] == 'slotNum':
+         Dispenser = dispenser.objects(id=request.POST.get('dispID'))[0]
+         compID = request.POST.get('compID')
+         Compartment = compartment.objects(id=compID)[0]
+         slotNum = int(request.POST.get('slotNum'))   
+            
+         Compartment.loaded = True
+         Compartment.save()
+         Dispenser.slots[slotNum] = Compartment
+         Dispenser.save()
+         params['message'] = "Loaded compartment into slot number " + str(slotNum + 1)
+         params['message_type'] = "success"
+
+         if len(free) == 1:
+            params['hide'] = True
+   return render(request, 'updatecompartment.html', params)
